requirements.txt
Django==4.0.1
django-ckeditor
djangorestframework
Pillow
django-cors-headers
